+++ src/main/java/com/keepgoing/keepserver/global/config/ExcelConfig.java
@@ -1,11 +1,11 @@
 package com.keepgoing.keepserver.global.config;
 
-import com.keepgoing.keepserver.domain.file.service.generate.GenerateExcelTemplate;
-import com.keepgoing.keepserver.domain.file.service.generate.GenerateExcelTemplateImpl;
-import com.keepgoing.keepserver.domain.file.service.mapper.RowMapper;
-import com.keepgoing.keepserver.domain.file.service.parser.ExcelParser;
-import com.keepgoing.keepserver.domain.file.service.parser.ExcelParserImpl;
-import com.keepgoing.keepserver.domain.file.service.validate.ExcelValidator;
+import com.keepgoing.keepserver.global.file.generate.GenerateExcelTemplate;
+import com.keepgoing.keepserver.global.file.generate.GenerateExcelTemplateImpl;
+import com.keepgoing.keepserver.global.file.RowMapper;
+import com.keepgoing.keepserver.global.file.ExcelParser;
+import com.keepgoing.keepserver.global.file.ExcelParserImpl;
+import com.keepgoing.keepserver.global.file.validate.ExcelValidator;
 import com.keepgoing.keepserver.domain.teacher.service.ItemDtoRowMapperImpl;
 import com.keepgoing.keepserver.domain.teacher.service.ItemExcelDtoValidator;
 import com.keepgoing.keepserver.domain.teacher.payload.ItemExcelDto;
