+++ src/main/java/com/keepgoing/keepserver/global/file/generate/ExcelGenerator.java
@@ -1,4 +1,4 @@
-package com.keepgoing.keepserver.domain.file.service.generate;
+package com.keepgoing.keepserver.global.file.generate;
 
 import com.keepgoing.keepserver.global.exception.excel.ExcelException;
 import org.apache.poi.ss.usermodel.Workbook;
