+++ src/main/java/com/keepgoing/keepserver/global/file/generate/GenerateExcelTemplate.java
@@ -1,4 +1,4 @@
-package com.keepgoing.keepserver.domain.file.service.generate;
+package com.keepgoing.keepserver.global.file.generate;
 
 import org.apache.poi.ss.usermodel.Workbook;
